Changed: update infos behaviour for healed sap and stamina
simple hack to pass different heal type to DeltaHP
show healing factor of sap and stamina like used by HP

note: in case we want to do this properly, and use a different rgba output color per type

edit EBehaviour located in game_share to add a new currative cast effect (eg. CAST_HEALHP, CAST_HEALSAP)
don't forget to edit stringToBehaviour() if the enum is modified, and insert a signed integer to CBehaviour (DeltaSap, DeltaSta)

client side we notice impulseCombatFlyingHpDelta() that introduce more or less what we want.
unfortunatly it is associated to the net callback "COMBAT:FLYING_HP_DELTA" which is not found anywhere else,
so i guess a dead end? any info about this?

back to client side, edit character_cl to copy the applyBehaviourFlyingHPs() function and method (DeltaHPColor etc).
recursively in the function applyBehaviour() insert the new currative cast effect (begin to end)
and replicate or modify the call to applyBehaviourFlyingHPs() who's passing a RGBA color to addHPOutput()
it is mainly here the change can be fine tuned and finalized.